mark-x-enhanced/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0

# Voice Processing
sounddevice==0.4.6
soundfile==0.12.1
vosk==0.3.44

# TTS - ElevenLabs
elevenlabs==0.2.27

# Wake Word Detection
pvporcupine==3.0.0

# LLM & AI
openai==1.12.0

# Database & Memory
sqlalchemy==2.0.25
alembic==1.13.1

# Vector Store & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Task Scheduling
APScheduler==3.10.4

# Async & Task Queue
asyncio
aiofiles==23.2.1

# Messaging Platform Integrations
python-telegram-bot==20.7
discord.py==2.3.2
pywhatkit==5.4

# Web Framework & API
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# HTTP Client
httpx==0.26.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
